my_learning/requirement.txt
Package               Version     
--------------------- ------------
amqp                  1.4.9       
antiorm               1.2.1       
anyjson               0.3.3       
asn1crypto            0.24.0      
attrs                 17.4.0      
Automat               0.6.0       
backcall              0.1.0       
baidu-aip             2.2.13.0    
billiard              3.3.0.23    
bleach                2.1.3       
bottle                0.12.16     
celery                3.1.26.post2
certifi               2018.1.18   
cffi                  1.12.2      
chardet               3.0.4       
click                 6.7         
colorama              0.3.7       
constantly            15.1.0      
cryptography          2.1.4       
cupshelpers           1.0         
cycler                0.10.0      
db                    0.1.1       
decorator             4.4.0       
Django                2.2         
django-celery         3.2.2       
django-rest-framework 0.1.0       
djangorestframework   3.9.2       
entrypoints           0.2.3.post1 
falcon                1.4.1       
Flask                 1.0.2       
html5lib              0.999999999 
httplib2              0.9.2       
hyperlink             17.3.1      
idna                  2.6         
incremental           16.10.1     
ipykernel             4.8.2       
ipython               7.4.0       
ipython-genutils      0.2.0       
ipywidgets            6.0.0       
itsdangerous          1.1.0       
jedi                  0.13.3      
Jinja2                2.10.1      
jsonschema            2.6.0       
jupyter-client        5.2.3       
jupyter-core          4.4.0       
keyring               10.6.0      
keyrings.alt          3.0         
kiwisolver            1.0.1       
kombu                 3.0.37      
MarkupSafe            1.1.1       
matplotlib            3.0.3       
mistune               0.8.3       
nbconvert             5.3.1       
nbformat              4.4.0       
netifaces             0.10.4      
notebook              5.4.1       
numpy                 1.16.2      
onboard               1.4.1       
opencv-python         4.1.0.25    
optionaldict          0.1.1       
pandas                0.24.2      
pandocfilters         1.4.2       
parso                 0.4.0       
pexpect               4.2.1       
pickleshare           0.7.5       
Pillow                5.1.0       
pip                   19.0.3      
prompt-toolkit        2.0.9       
pyasn1                0.4.2       
pyasn1-modules        0.2.4       
pycairo               1.16.2      
pycparser             2.19        
pycrypto              2.6.1       
pycups                1.9.73      
Pygments              2.3.1       
pygobject             3.28.2      
PyMySQL               0.9.3       
pyOpenSSL             17.5.0      
pyparsing             2.4.0       
pyserial              3.4         
pysmbc                1.0.15.6    
python-dateutil       2.8.0       
python-mimeparse      1.6.0       
pytz                  2019.1      
pyxdg                 0.25        
PyYAML                3.12        
pyzmq                 17.0.0      
reportlab             3.4.0       
requests              2.18.4      
SecretStorage         2.3.1       
Send2Trash            1.4.2       
service-identity      16.0.0      
setuptools            41.0.0      
simplegeneric         0.8.1       
six                   1.11.0      
sqlparse              0.3.0       
terminado             0.8.1       
testpath              0.3.1       
tornado               6.0.2       
traitlets             4.3.2       
Twisted               17.9.0      
urllib3               1.22        
utils                 0.9.0       
vine                  1.3.0       
wcwidth               0.1.7       
webencodings          0.5         
wechatpy              1.8.1       
Werkzeug              0.15.2      
WeRoBot               1.9.0       
wheel                 0.33.1      
xmltodict             0.12.0      
zope.interface        4.3.2       
